Move Enums to Content.Shared.Arcade.Enums namespace
