+++ Content.Shared/Arcade/Components/BlockGameArcadeComponent.cs
@@ -1,3 +1,4 @@
+using Content.Shared.Arcade.Enums;
 using Content.Shared.Arcade.Systems;
 using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
