+++ Content.Shared/Arcade/Enums/ArcadeUiKey.cs
@@ -1,6 +1,6 @@
 using Robust.Shared.Serialization;
 
-namespace Content.Shared.Arcade;
+namespace Content.Shared.Arcade.Enums;
 
 [Serializable, NetSerializable]
 public enum ArcadeUiKey : byte
