+++ Content.Shared/Arcade/Enums/SpaceVillainWireStatus.cs
@@ -1,6 +1,6 @@
 using Robust.Shared.Serialization;
 
-namespace Content.Shared.Arcade;
+namespace Content.Shared.Arcade.Enums;
 
 [Serializable, NetSerializable]
 public enum SpaceVillainArcadeWireStatus
